raspberryPi/requirements.txt
youtube_dl
python-socketio[asyncio_client]
pygame
